Clean up some generated stuff + general clean up
