+++ packages/components/Shadow/jest.config.js
@@ -1,2 +1,2 @@
 const { configureReactNativeJest } = require('@fluentui-react-native/scripts');
-// module.exports = configureReactNativeJest('android');
+module.exports = configureReactNativeJest('android');
