+++ packages/components/Shadow/src/Shadow.types.ts
@@ -1,22 +1,13 @@
 import { ViewProps } from 'react-native';
-import { TextProps } from '@fluentui-react-native/experimental-text';
 import { IBorderTokens, IColorTokens, LayoutTokens } from '@fluentui-react-native/tokens';
 
 export const shadow = 'Shadow';
 
 export type ShadowDepth = '2' | '4' | '8' | '16' | '28' | '64';
 
-export interface ShadowTokens extends LayoutTokens, IBorderTokens, IColorTokens {
-  small?: ShadowTokens;
-  medium?: ShadowTokens;
-  large?: ShadowTokens;
-}
+export interface ShadowTokens extends LayoutTokens, IBorderTokens, IColorTokens {}
 
-export interface ShadowProps {
-  depth?: ShadowDepth;
-}
-
-export interface ShadowProps extends ViewProps, ViewStyle, IBorderTokens {
+export interface ShadowProps extends ViewProps, IBorderTokens {
   depth?: ShadowDepth;
 }
